# docs/swagger.yaml
basePath: /
definitions:
...
  admin.createJobRequest:
    properties:
      command:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      image:
        type: string
      name:
        type: string
      priority:
        type: integer
      time_limit:
        type: integer
      user_id:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for Stream application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stream API
  version: "1.0"
paths:
...
  /admin/agents:
    get:
      description: Returns currently connected render agents and current runtime stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List connected render agents
      tags:
      - admin-render
  /admin/agents/{id}/restart:
    post:
      description: Sends a restart command to a currently connected render agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Restart connected render agent
      tags:
      - admin-render
  /admin/agents/{id}/update:
    post:
      description: Sends an update command to a currently connected render agent
      parameters:
      - description: Agent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update connected render agent
      tags:
      - admin-render
...
  /admin/jobs:
    get:
      description: Returns paginated render jobs for admin management
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Agent ID
        in: query
        name: agent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List render jobs
      tags:
      - admin-render
    post:
      consumes:
      - application/json
      description: Queues a new render job for agents
      parameters:
      - description: Job payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/admin.createJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create render job
      tags:
      - admin-render
  /admin/jobs/{id}:
    get:
      description: Returns a render job by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get render job detail
      tags:
      - admin-render
  /admin/jobs/{id}/cancel:
    post:
      description: Cancels a pending or running render job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Cancel render job
      tags:
      - admin-render
  /admin/jobs/{id}/logs:
    get:
      description: Returns plain text logs for a render job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get render job logs
      tags:
      - admin-render
  /admin/jobs/{id}/retry:
    post:
      description: Retries a failed or cancelled render job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Retry render job
      tags:
      - admin-render
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"